+++ geo/benches/area.rs
@@ -11,9 +11,7 @@ fn criterion_benchmark(c: &mut Criterion) {
         let polygon = Polygon::new(LineString::<f32>::from(points), vec![]);
 
         bencher.iter(|| {
-            criterion::black_box(|| {
-                polygon.signed_area();
-            });
+            criterion::black_box(criterion::black_box(&polygon).signed_area());
         });
     });
 }