+++ geo/benches/simplify.rs
@@ -12,7 +12,9 @@ fn criterion_benchmark(c: &mut Criterion) {
         let points = include!("../src/test_fixtures/louisiana.rs");
         let ls: LineString<f32> = points.into();
         bencher.iter(|| {
-            let _ = ls.simplifyvw(&0.0005);
+            criterion::black_box(
+                criterion::black_box(&ls).simplifyvw(criterion::black_box(&0.0005)),
+            );
         });
     });
 
@@ -20,7 +22,9 @@ fn criterion_benchmark(c: &mut Criterion) {
         let points = include!("../src/test_fixtures/louisiana.rs");
         let ls: LineString<f64> = points.into();
         bencher.iter(|| {
-            let _ = ls.simplifyvw(&0.0005);
+            criterion::black_box(
+                criterion::black_box(&ls).simplifyvw(criterion::black_box(&0.0005)),
+            );
         });
     });
 
@@ -28,7 +32,9 @@ fn criterion_benchmark(c: &mut Criterion) {
         let points = include!("../src/test_fixtures/louisiana.rs");
         let ls: LineString<f32> = points.into();
         bencher.iter(|| {
-            let _ = ls.simplifyvw_preserve(&0.0005);
+            criterion::black_box(
+                criterion::black_box(&ls).simplifyvw_preserve(criterion::black_box(&0.0005)),
+            );
         });
     });
 
@@ -36,7 +42,9 @@ fn criterion_benchmark(c: &mut Criterion) {
         let points = include!("../src/test_fixtures/louisiana.rs");
         let ls: LineString<f32> = points.into();
         bencher.iter(|| {
-            let _ = ls.simplifyvw_preserve(&0.0005);
+            criterion::black_box(
+                criterion::black_box(&ls).simplifyvw_preserve(criterion::black_box(&0.0005)),
+            );
         });
     });
 }